fix: Remove all warnings and fix compilation errors
- Fixed enum variant naming conventions (SCREAMING_CASE to CamelCase) throughout codebase
- Updated all SyntaxKind enum variants to follow Rust naming conventions
- Fixed unused variable warnings by prefixing with underscore
- Fixed dead code warnings with appropriate attributes
- Fixed borrow checker error in skalp-incremental
- Fixed module resolution error in skalp-formal (skalp_mir -> skalp_lir)
- Fixed duplicate error variant in skalp-backends (removed Io, kept IoError)
- Fixed syntax errors in asic.rs
- Applied cargo fix suggestions for automatic fixes
- Main package now compiles without warnings

Some crates still have compilation errors but main functionality is restored.

+++ crates/skalp-formal/src/bmc.rs
@@ -3,7 +3,7 @@
 use crate::smt::{SmtSolver, SatResult};
 use crate::property::{Property, TemporalFormula};
 use crate::{FormalResult, FormalError, PropertyStatus};
-use skalp_mir::lir::LirDesign;
+use skalp_lir::LirDesign;
 
 /// Bounded Model Checker
 pub struct BoundedModelChecker {
@@ -48,7 +48,7 @@ impl BoundedModelChecker {
 
             // Unroll transition relation
             for i in 0..k {
-                let transition = ts.get_transition(i);
+                let transition = ts.get_transition(i as usize);
                 self.solver.assert(&transition);
             }
